creating an extended version of Garden to complete Part Two but having severe performance issues
