+++ src/components/Days/DayFive/partTwo.ts
@@ -1,1 +1,6 @@
-export const partTwoHandler = (data: string) => data.length;
+import { AdvancedGarden } from "./AdvancedGarden";
+
+export const partTwoHandler = (data: string) => {
+  const garden = new AdvancedGarden(data);
+  return garden.lowestLocation;
+};
